First Attempt for Class Generation on Persistence Project
