+++ Contractor.Core/Projects/Persistence/Generation/DTOs/DbEntityDetail/DbEntityDetailGeneration.cs
@@ -1,0 +1,60 @@
+﻿using Contractor.Core.Options;
+using Contractor.Core.Tools;
+using System.IO;
+
+namespace Contractor.Core.Projects.Persistence
+{
+    internal class DbEntityDetailGeneration : ClassGeneration
+    {
+        private static readonly string TemplatePath =
+            Path.Combine(PersistenceProjectGeneration.TemplateFolder, "DbEntityDetailTemplate.txt");
+
+        private static readonly string FileName = "DbEntityDetail.cs";
+
+        private readonly DbDtoDetailMethodsAddition dbDtoDetailMethodsAddition;
+        private readonly DbDtoDetailFromMethodsAddition dbDtoDetailFromMethodsAddition;
+        private readonly DbDtoDetailToMethodsAddition dbDtoDetailToMethodsAddition;
+        private readonly DtoAddition dtoAddition;
+        private readonly DtoPropertyAddition propertyAddition;
+
+        public DbEntityDetailGeneration()
+        {
+        }
+
+        protected override void AddDomain(IDomainAdditionOptions options)
+        {
+        }
+
+        protected override void AddEntity(IEntityAdditionOptions options)
+        {
+            this.dtoAddition.AddDto(options, PersistenceProjectGeneration.DomainFolder, TemplatePath, FileName);
+        }
+
+        protected override void AddProperty(IPropertyAdditionOptions options)
+        {
+            this.propertyAddition.AddPropertyToDTO(options, PersistenceProjectGeneration.DomainFolder, FileName);
+            this.dbDtoDetailMethodsAddition.Add(options, PersistenceProjectGeneration.DomainFolder, FileName);
+        }
+
+        protected override void Add1ToNRelation(IRelationAdditionOptions options)
+        {
+            // From
+            IPropertyAdditionOptions optionsFrom = 
+                RelationAdditionOptions.GetPropertyForFrom(options, $"IEnumerable<IDb{options.EntityNameTo}>", options.EntityNamePluralTo);
+            this.propertyAddition.AddPropertyToDTO(optionsFrom, PersistenceProjectGeneration.DomainFolder, FileName,
+                $"{options.ProjectName}.Contract.Persistence.Modules.{options.DomainTo}.{options.EntityNamePluralTo}");
+
+            this.dbDtoDetailFromMethodsAddition.Add(options, PersistenceProjectGeneration.DomainFolder, FileName,
+                $"{options.ProjectName}.Persistence.Modules.{options.DomainTo}.{options.EntityNamePluralTo}");
+
+            // To
+            IPropertyAdditionOptions optionsTo = 
+                RelationAdditionOptions.GetPropertyForTo(options, $"IDb{options.EntityNameFrom}", options.EntityNameFrom);
+            this.propertyAddition.AddPropertyToDTO(optionsTo, PersistenceProjectGeneration.DomainFolder, FileName,
+                $"{options.ProjectName}.Contract.Persistence.Modules.{options.DomainFrom}.{options.EntityNamePluralFrom}");
+
+            this.dbDtoDetailToMethodsAddition.Add(options, PersistenceProjectGeneration.DomainFolder, FileName,
+                $"{options.ProjectName}.Persistence.Modules.{options.DomainFrom}.{options.EntityNamePluralFrom}");
+        }
+    }
+}
