+++ Contractor.Core/Projects/Persistence/Generation/DTOs/DbEntityDetail/Services/DbDtoDetailFromMethodsAddition.cs
@@ -3,30 +3,31 @@ using Contractor.Core.Options;
 using Contractor.Core.Tools;
 using System.IO;
 
-namespace Contractor.Core.Projects
+namespace Contractor.Core.Projects.Persistence
 {
-    internal class DbDtoDetailToMethodsAddition
+    internal class DbDtoDetailFromMethodsAddition
     {
         public PathService pathService;
 
-        public DbDtoDetailToMethodsAddition(PathService pathService)
+        public DbDtoDetailFromMethodsAddition(PathService pathService)
         {
             this.pathService = pathService;
         }
 
-        public void Add(IRelationAdditionOptions options, string domainFolder, string templateFileName, string namesapceToAdd)
+        public void Add(IRelationAdditionOptions options, string domainFolder, string templateFileName, string namespaceToAdd)
         {
             string filePath = GetFilePath(options, domainFolder, templateFileName);
             string fileData = UpdateFileData(options, filePath);
 
-            fileData = UsingStatements.Add(fileData, namesapceToAdd);
+            fileData = UsingStatements.Add(fileData, "System.Linq");
+            fileData = UsingStatements.Add(fileData, namespaceToAdd);
 
             CsharpClassWriter.Write(filePath, fileData);
         }
 
         private string GetFilePath(IRelationAdditionOptions options, string domainFolder, string templateFileName)
         {
-            IEntityAdditionOptions entityOptions = RelationAdditionOptions.GetPropertyForTo(options);
+            IEntityAdditionOptions entityOptions = RelationAdditionOptions.GetPropertyForFrom(options);
             string absolutePathForDTOs = this.pathService.GetAbsolutePathForDTOs(entityOptions, domainFolder);
             string fileName = templateFileName.Replace("Entity", entityOptions.EntityName);
             string filePath = Path.Combine(absolutePathForDTOs, fileName);
@@ -39,10 +40,11 @@ namespace Contractor.Core.Projects
 
             // ----------- DbSet -----------
             StringEditor stringEditor = new StringEditor(fileData);
-            stringEditor.NextThatContains("FromEf" + options.EntityNameTo);
+            stringEditor.NextThatContains("FromEf" + options.EntityNameFrom);
             stringEditor.Next(line => line.Trim().Equals("};"));
 
-            stringEditor.InsertLine($"                {options.EntityNameFrom} = Db{options.EntityNameFrom}.FromEf{options.EntityNameFrom}(ef{options.EntityNameTo}.{options.EntityNameFrom}),");
+            stringEditor.InsertLine($"                {options.EntityNamePluralTo} = ef{options.EntityNameFrom}.{options.EntityNamePluralTo}" +
+                $".Select(ef{options.EntityNameTo} => Db{options.EntityNameTo}.FromEf{options.EntityNameTo}(ef{options.EntityNameTo})),");
 
             return stringEditor.GetText();
         }
