+++ Contractor.Core/Projects/Persistence/Generation/DTOs/EfEntity/Services/EfDtoContructorHashSetAddition.cs
@@ -3,7 +3,7 @@ using Contractor.Core.Options;
 using Contractor.Core.Tools;
 using System.IO;
 
-namespace Contractor.Core.Projects
+namespace Contractor.Core.Projects.Persistence
 {
     internal class EfDtoContructorHashSetAddition
     {
